+++ crates/tsz-emitter/src/transforms/async_es5_ir.rs
@@ -81,6 +81,8 @@ mod hoists;
 mod loop_control;
 #[path = "async_es5_ir_names.rs"]
 mod names;
+#[path = "async_es5_ir_rest_params.rs"]
+mod rest_params;
 #[path = "async_es5_ir_state.rs"]
 mod state;
 #[path = "async_es5_ir_statement_helpers.rs"]
@@ -503,81 +505,6 @@ impl<'a> AsyncES5Transformer<'a> {
                 body_source_range: None,
             }
         }
-    }
-
-    /// Return `(name, index)` of a trailing identifier rest parameter, if any.
-    /// Binding-pattern rest parameters are left to the existing path.
-    fn identifier_rest_param_info(
-        &self,
-        params: &tsz_parser::parser::NodeList,
-    ) -> Option<(String, usize)> {
-        for (index, &param_idx) in params.nodes.iter().enumerate() {
-            let param_node = self.arena.get(param_idx)?;
-            let Some(param) = self.arena.get_parameter(param_node) else {
-                continue;
-            };
-            if !param.dot_dot_dot_token {
-                continue;
-            }
-            let name_node = self.arena.get(param.name)?;
-            if name_node.kind != tsz_scanner::SyntaxKind::Identifier as u16 {
-                return None;
-            }
-            let name =
-                crate::transforms::emit_utils::identifier_text_or_empty(self.arena, param.name);
-            if name.is_empty() {
-                return None;
-            }
-            return Some((name, index));
-        }
-        None
-    }
-
-    /// Emit `var <rest> = []; for (<idx> = N; <idx> < arguments.length; <idx>++)
-    /// { <rest>[<idx> - N] = arguments[<idx>]; }` for an ES5 rest parameter.
-    fn push_rest_param_prologue(
-        &self,
-        body: &mut Vec<IRNode>,
-        rest_name: &str,
-        rest_index: usize,
-        index_name: &str,
-    ) {
-        body.push(IRNode::VarDecl {
-            name: rest_name.to_string().into(),
-            initializer: Some(Box::new(IRNode::ArrayLiteral(Vec::new()))),
-        });
-        let lhs_index = if rest_index > 0 {
-            IRNode::binary(
-                IRNode::id(index_name.to_string()),
-                "-",
-                IRNode::number(rest_index.to_string()),
-            )
-        } else {
-            IRNode::id(index_name.to_string())
-        };
-        let copy = Self::expression_statement(IRNode::assign(
-            IRNode::elem(IRNode::id(rest_name.to_string()), lhs_index),
-            IRNode::elem(
-                IRNode::id("arguments".to_string()),
-                IRNode::id(index_name.to_string()),
-            ),
-        ));
-        body.push(IRNode::ForStatement {
-            initializer: Some(Box::new(IRNode::assign(
-                IRNode::id(index_name.to_string()),
-                IRNode::number(rest_index.to_string()),
-            ))),
-            condition: Some(Box::new(IRNode::binary(
-                IRNode::id(index_name.to_string()),
-                "<",
-                IRNode::prop(IRNode::id("arguments".to_string()), "length"),
-            ))),
-            incrementor: Some(Box::new(IRNode::PostfixUnaryExpr {
-                operand: Box::new(IRNode::id(index_name.to_string())),
-                operator: "++".into(),
-            })),
-            body: Box::new(IRNode::Block(vec![copy])),
-        });
     }
 
     pub fn transform_async_generator_inner_function(
